LASP Data Filter Updates
Multiple updates to main codes to implement the filtered LASP irradiance data to use as input for the radiation pressure force function, as well as some more testing for plots of 2D and 3D data.

New files:
bulkvelcomparison.py: Created to plot two lines of the bulk velocities over the solar cycle to compare with the Izmodenov et al. 2013 paper.

parallelpstraj_supplement.py: A new file to re-run points that through the odeint warning in the 3D code, so they can be rerun at a different time resolution